Scores stored in db

+++ ALIAGA_SERGIO_PALMA_PEREZ_PRACTICA_ANDROID_M08UF1/app/src/main/java/edu/fje/aliaga_sergio_palma_perez_practica_android_m08uf1/Score.java
@@ -1,7 +1,7 @@
 package edu.fje.aliaga_sergio_palma_perez_practica_android_m08uf1;
 
 public class Score {
-    static public int time;
-    static public int points;
+    static public long time;
+    static public long points;
     static public int level;
 }
